requirements.txt
# Web Framework
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Image Processing
Pillow>=10.2.0,<11.0.0
opencv-python>=4.8.0
numpy>=1.24.0,<2.0.0

# Deep Learning (adjust based on your model)
torch>=2.0.0
torchvision>=0.15.0
# tensorflow>=2.15.0  # Uncomment if using TensorFlow
ultralytics>=8.0.0  # YOLOv8 for object detection

# Database for real-time sync
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0

# Google Cloud (for deployment)
google-cloud-storage>=2.10.0
google-cloud-compute>=1.10.0

